adding namespace
this is just a start and examples and syntax would need to be adjusted, but it really would make sense to use namespaces for the XML syntax. it also would be necessary to specify the processing model: what to do when encountering a message with unknown elements in the UBER namespace, with unknown attributes, and elements from non-UBER namespaces. this will improve robustness in the long run. it also should be part of an evolution model: how will UBER evolve, how will it be able to co-exist with other vocabularies?